+++ CHANGELOG.md
@@ -1,2 +1,9 @@
-## [0.1.0] 
+## [0.1.2]
+
+* Added:
+* menuOverlapWidth - width of the menu overlay above the body. This setting is useful for menu decoration (Shifts the shadow and scrim layer).
+* backgroundColor - background color under the menu and body.
+
+## [0.1.0]
+
 * First release